[update] 途中結果をとるようにした
